src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
autopep8==1.7.0
    # via -r requirements.in
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
django==4.1.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
idna==3.4
    # via requests
psycopg2-binary==2.9.3
    # via -r requirements.in
pycodestyle==2.9.1
    # via autopep8
pyjwt==2.5.0
    # via djangorestframework-simplejwt
pytz==2022.2.1
    # via djangorestframework
requests==2.28.1
    # via -r requirements.in
sqlparse==0.4.3
    # via django
toml==0.10.2
    # via autopep8
tzdata==2022.4
    # via django
urllib3==1.26.12
    # via requests
